Move min quantity update page to a modal
