+++ src/Components/Facility/MinQuantityList.tsx
@@ -4,8 +4,10 @@ import loadable from "@loadable/component";
 import { statusType, useAbortableEffect } from "../../Common/utils";
 import { getMinQuantity, getAnyFacility } from "../../Redux/actions";
 import Pagination from "../Common/Pagination";
-import { Link, navigate } from "raviger";
+import { navigate } from "raviger";
 import { RoleButton } from "../Common/RoleButton";
+import { MinQuantityRequiredModal } from "./MinQuantityRequiredModal";
+import { Button } from "@material-ui/core";
 const Loading = loadable(() => import("../Common/Loading"));
 const PageTitle = loadable(() => import("../Common/PageTitle"));
 
@@ -20,6 +22,9 @@ export default function MinQuantityList(props: any) {
   const [currentPage, setCurrentPage] = useState(1);
   const [totalCount, setTotalCount] = useState(0);
   const [facilityName, setFacilityName] = useState("");
+  const [showMinQuantityRequiredModal, setShowMinQuantityRequiredModal] =
+    useState(false);
+  const [selectedItem, setSelectedItem] = useState({ id: 0, item_id: 0 });
   const limit = 14;
 
   const fetchData = useCallback(
@@ -75,8 +80,10 @@ export default function MinQuantityList(props: any) {
               {inventoryItem.item_object?.name}
             </p>
           </div>
-          <Link
-            href={`/facility/${facilityId}/inventory/${inventoryItem.id}/update/${inventoryItem.item_object?.id}`}
+          <Button
+            onClick={() => {
+              alert(3);
+            }}
           >
             <div className="sm:hidden flex justify-between items-center w-full">
               <div className="flex flex-col">
@@ -106,7 +113,7 @@ export default function MinQuantityList(props: any) {
                 </svg>
               </div>
             </div>
-          </Link>
+          </Button>
         </td>
         <td className="px-5 py-5 border-b border-gray-200 text-sm sm:flex hidden w-full justify-between">
           <p className="text-gray-900 whitespace-nowrap lowercase mt-2">
@@ -120,11 +127,13 @@ export default function MinQuantityList(props: any) {
               color: "primary",
               size: "medium",
             }}
-            handleClickCB={() =>
-              navigate(
-                `/facility/${facilityId}/inventory/${inventoryItem.id}/update/${inventoryItem.item_object?.id}`
-              )
-            }
+            handleClickCB={() => {
+              setSelectedItem({
+                id: inventoryItem.id,
+                item_id: inventoryItem.item_object?.id,
+              });
+              setShowMinQuantityRequiredModal(true);
+            }}
             disableFor="readOnly"
             buttonType="materialUI"
           >
@@ -138,7 +147,7 @@ export default function MinQuantityList(props: any) {
       <tr className="bg-white">
         <td
           colSpan={3}
-          className="px-5 py-5 border-b border-gray-200 text-center"
+          className="pxf-5 py-5 border-b border-gray-200 text-center"
         >
           <p className="text-gray-500 whitespace-nowrap">
             No item with minimum quantity set
@@ -227,6 +236,19 @@ export default function MinQuantityList(props: any) {
           {inventoryItem}
         </div>
       </div>
+      {showMinQuantityRequiredModal && (
+        <MinQuantityRequiredModal
+          facilityId={facilityId}
+          inventoryId={selectedItem.id}
+          itemId={selectedItem.item_id}
+          show={showMinQuantityRequiredModal}
+          handleClose={() => setShowMinQuantityRequiredModal(false)}
+          handleUpdate={() => {
+            fetchData({ aborted: false });
+            setShowMinQuantityRequiredModal(false);
+          }}
+        />
+      )}
     </div>
   );
 }
