+++ src/Router/AppRouter.tsx
@@ -25,7 +25,6 @@ import InventoryLog from "../Components/Facility/InventoryLog";
 import { AddInventoryForm } from "../Components/Facility/AddInventoryForm";
 import { SetInventoryForm } from "../Components/Facility/SetInventoryForm";
 import MinQuantityList from "../Components/Facility/MinQuantityList";
-import { UpdateMinQuantity } from "../Components/Facility/UpdateMinQuantity";
 import { ShiftCreate } from "../Components/Patient/ShiftCreate";
 import UserProfile from "../Components/Users/UserProfile";
 import ShiftBoardView from "../Components/Shifting/BoardView";
@@ -307,17 +306,6 @@ const routes = {
     facilityId,
     inventoryId,
   }: any) => <InventoryLog facilityId={facilityId} inventoryId={inventoryId} />,
-  "/facility/:facilityId/inventory/:inventoryId/update/:itemId": ({
-    facilityId,
-    inventoryId,
-    itemId,
-  }: any) => (
-    <UpdateMinQuantity
-      facilityId={facilityId}
-      inventoryId={inventoryId}
-      itemId={itemId}
-    />
-  ),
   "/facility/:facilityId/assets/new": ({ facilityId }: any) => (
     <AssetCreate facilityId={facilityId} />
   ),
